requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
fastapi==0.115.11
filetype==1.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langgraph==0.3.14
langgraph-checkpoint==2.0.21
langgraph-checkpoint-postgres==2.0.18
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
msgpack==1.1.0
numpy==2.2.4
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
proto-plus==1.26.1
protobuf==5.29.3
psycopg==3.2.6
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
